pyproject.toml
[tool.poetry]
name = "doc-ocr"
version = "0.1.0"
description = ""
authors = ["ekusiadadus <ekusiadadus@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.4"
opencv-python = "^4.6.0"
Pillow = "^9.3.0"
google-cloud-vision = "^3.1.4"
reportlab = "^3.6.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
